strongly typed error codes
Summary:
Currently, serverless-IDE puts a lot of error messages into the VSCode output window for every message it can't handle due to being stopped or initializing.

But these should not be shown to the user. We already changed clientLsp to use RequestCancelled error code for messages declined due to these reasons, since that's the only error code that doesn't show up in the output window.

I want to do the same for serverless-IDE.

This diff is a first step. It changes to strongly type the error codes.

Reviewed By: arxanas

Differential Revision: D19800840

fbshipit-source-id: 6943a9aa33b1043606c57b3ededed4ef35430064