Avoid double-loads for type checks
Summary:
For specialized types, CheckLoc  / CheckStack / CheckMBase have to load the value field, but we don't have the vasm-level memory tracking needed to avoid another load when we actually get to the LdLoc / LdStk / LdMem. We can easily avoid the extra load in irgen, however, by splitting up the guard into an in-frame check of the base type, a load, and a CheckType for the specialization. If we don't use the value, the final code-gen is unchanged by this change, but if we do use the value, we save one load.

We shouldn't do this optimization while forming a region because we associate GuardConstraints with the actual CheckLoc instruction.

This change has a minimal impact right now: only about 1% of guards at the beginning of a region are to DataTypeSpecialized arrays or objects. (Most specialized guards are internal to a region, and we usually end up doing a CheckType for those since we've already materialized the value.) However, it helps with the array-like virtualization project; I measured about a 0.5% CPU regression without this change and a 0.2% regression with it.

Reviewed By: ricklavoie

Differential Revision: D19688873

fbshipit-source-id: b0e894b938d5af81b4cbbcfdc79aeac560a27e64