Support recursive provenance tagging
Summary:
This function is tricky to write because we want to avoid a) backtracing and b) copying the array unless it or its children need a tag, and if we do start changing array values, we only want to backtrace once. There are a few subtleties to watch out for here:
1. We don't want to trace through object properties, so if we find one, we'll raise a notice. We'll raise at most one notice per call to tag_provenance_here. We keep a state to enforce this bound.
2. ArrayIter normally inc-refs the base to keep it from being mutated. We use "local iteration", which meant for loops that mutate an array in place. Local iteration looks quite different from other iterators.
3. When we are forced to COW any array, we must COW all of its children, even if they have a refcount of 1. We handle this by passing a "cow" parameter down the recursion stack. (Unlike the fields in RecursionState, edits to "cow" should not be broadcast up the stack.)

I added a few tests exercising different edge cases here. The big one just checks us tagging a variety of varrays and darrays, through dicts. There's a test specifically for a value that should get COWed - the test doesn't confirm that it does, but it confirms that if it did, we didn't make refcounting mistakes. There's also a test for the recursive COW behavior that includes a refcount 1 array within a refcount 2 array.

Reviewed By: jjgriego

Differential Revision: D19844499

fbshipit-source-id: 51718d523663c226e05552287e4e3afde270f7a5