Properly handle dynamic for array access
Summary:
When we early solve for the `[]` operator, we widen the type. If a type cannot be widened then it defaults to nothing. Thus if you attempt to widen a type such as `~Vector<#1>` we will end up with `nothing | Vector<_>` which is not a widen version of the original type. The change to fix this is simple, we widen `dynamic` to `dynamic`, so in the above case we will properly widen the type to `dynamic | Vector<_>`.

This fixes a few discrepancies under like types.

Reviewed By: manzyuk

Differential Revision: D19815800

fbshipit-source-id: 22607d25c6a8b27c120f94f348096e3304fa2578