Generate an SSwitch for 86cinits
Summary:
We've always just generated a long list of if/elses, but since we
control the function, and the calls to the function we can just
generate an SSwitch instead. For classes with lots of constants this
is much more efficient both for hhbbc to optimize away, and for hhvm
at runtime (in non-repo mode, or when hhbbc failed to optimize the
switch away).

Reviewed By: oulgen, alexeyt

Differential Revision: D19857845

fbshipit-source-id: bf57f4fc745836691190d5471275c501223c25ed