fix lint HackLint5624 w.r.t PU dependent types
Summary:
PU might use type parameters as dependent types, which currently makes
the linter fires for a wrong reason: we need the name in the return type
to express the type dependency on w.r.t. to the input `$x`.

This diff fixes the linter w.r.t to PU dependent types.

Reviewed By: manzyuk

Differential Revision: D19856209

fbshipit-source-id: 0f7826fe13bb88b5363710bf9e4fc2fd26dee24c