remove GLobal_naming_options.get in serverTypeCheck
Summary:
We used to sometimes check whether shallow decls are enabled via Global_naming_options.get, and sometimes via the tcopt/ctx that was passed down. I had encountered a bug where the two differed. Hence this diff, part of a stack to make the passed-down tcopt/ctx be the sole source of truth about whether we're using shallow decls.

The goal of Global_naming_options was to protect the integrity of the decl store. This is still an important goal. It's not clear what is the best way to do this is.

Reviewed By: arxanas

Differential Revision: D19826766

fbshipit-source-id: 219ab2b43463836649302ab434dbc909808543a0