switch decl_fun_utils from Global_naming_options to ctx
Summary:
This diff is interesting!

This single line is the only place in the codebase that uses Global_naming_options for something important.

(The only other places were ones which used it for the shallow_decl flag. I switched them over to ctx, since that made more sense for the shallow-decl question).

As for this last remaining place, it's unclear. Should this place be using Global_naming_options or ctx? Which one is correct to protect the integrity of the decl store?

I kind of want to switch over this SOLE user of Global_naming_options.get. It just doesn't feel right to have a big solid general machinery which is used in only one place.

Reviewed By: arxanas

Differential Revision: D19827216

fbshipit-source-id: bd9ac6e5ecbdc9185dc85500d2ece68ebb0d9ea9