remove Global_naming_options.get in decl_compare
Summary:
We used to sometimes check whether shallow decls are enabled via Global_naming_options.get, and sometimes via the tcopt/ctx that was passed down. I had encountered a bug where the two differed. Hence this diff, part of a stack to make the passed-down tcopt/ctx be the sole source of truth about whether we're using shallow decls.

The goal of Global_naming_options was to protect the integrity of the decl store. This is still an important goal. It's not clear what is the best way to do this is.

Reviewed By: arxanas

Differential Revision: D19826967

fbshipit-source-id: 8d9f5344d83aae3875033636a593cdd21b50e368