Rewrite type parameter / atom check for members
Summary:
During typechecking of PU dependent types `Class:Enum:Member:T`, the
`Member` part can be either a type parameter or a ground atom name.
Until now, the check was partly done during localization, and subtyping.
But since both are mutually dependent, this lead to cases where good code
was rejected.
In the current implementation, the detection is shifted to naming/localization:
- just after parsing, each member in the ast is tagged with an `Unknown` flag
- during naming, we can check if this member is a bound type parameter, or
  something else. In the former case, we flag it as `TypeParameter`. Otherwise
  we leave it `Unknown`.
- during localization, if it is still `Unknown`, we can now check if it is an
  atom, in which case we flag it as `Atom`. Otherwise, we can safely report an
  issue as the member is not bound anywhere.

If a type parameter is shadowing an atom name, we consider the member to be a
type parameter, as it is the most recent binding.

Reviewed By: manzyuk

Differential Revision: D19853266

fbshipit-source-id: ec7dfda8138e6081cf82787233bab7fe11d43de7