Fix opt_array_mark_legacy
Summary:
These should have had the catch block installed;

I don't think it is a significant issue since we don't emit the optimizations when irgen-ing a `NativeImpl` and that's the circumstance under which we will assert without the additional magic that make_opt_catch provides.

We do neglect to decref the params without before this patch though, which should be addressed.

Differential Revision: D19751269

fbshipit-source-id: 90efe5e7e9039e2ba5106f9847d5079303c13105