HHVM Debugger: variablesResponse should always include a variablesReference
Summary: A recent change to VS Code is more strictly enforcing the response to a variables request, and we are currently omitting variablesReference for primitive values with no childrne. Per the DAP protocol spec, this should always be present and should actually be 0 instead of undefined in these cases. This change was introduced in https://github.com/microsoft/vscode/commit/d42882c2260f511339b4e8aede225d2d3097f1f3

Reviewed By: alexeyt, solomatov

Differential Revision: D19907896

fbshipit-source-id: 4a8a4294e1bc2c6c28e1cfa9cd29a9ad03c1ad8b