Return valid SSATmp even when on unreachable path
Summary:
In an unlikely situation on an unreachable path, we could `ldLocWarn`
a local that has been proven to never warn by HHBBC.  This generated an
Unreachable bytecode, but did not return a valid SSATmp.  This caused future
compilation to fail.  Instead we now return a valid TBottom to keep things
going.

Reviewed By: oulgen

Differential Revision: D19925878

fbshipit-source-id: 6b0b0413c97c23a30aed4bd241a654033bc1d759