Comprehensive unreachable code fix
Summary:
After generating a control flow op in HHIR, we assume that we'll fall through except when we use special control-flow primitives like isThen, cond, etc. As a result, if we emit a terminal IR op, regular fallthrough codegen is all unreachable. This terminal op may be an Unreachable op, a throw op, or a type-guard that got simplified to a Jmp to taken; the logic works regardless. We can take advantage of this knowledge in many places to drastically simplify our initial generated IR. At a macro level, in region-tracelet and translate-region, we can break tracelets and end region blocks early once we're in an unreachable state. At a micro level, we never need to append new IR opcodes once we're in this state.

This diff adds an inUnreachableState() method to IRBuilder and uses it in these places. Just doing so naively triggers a few asserts, though, because we don't cut the per-bytecode emit function short when we get into these states. I patched two of these places:
1. In irgen for static properties, we made assertions about the FrameState that aren't true when we reach unreachable code.
2. irgen for ContKey and ContCurrent always resulted in a side exit, because it loaded an object and then checked if it was null.

Strictly speaking, I didn't need to patch item 2 because breaking the tracelet there achieves the same result. But I opted to do so, because why keep around dead irgen?

It's a bit strange that we'll keep trying to emit more IR for a single bytecode after getting into an unreachable state, but I think it's not too bad. FrameState will still track the stack offset, which is why popC / push / etc. functions continue working in these states. Since we provide a TBottom SSATmp result for these ops, we can continue to thread it around and do type-checks. On the whole, the situation isn't too different from what we do today, except that we detect unreachable code significantly earlier and we stop irgen as soon as we're done with that bytecode. See also D19891167 for a smaller-scale fix.

Reviewed By: ricklavoie

Differential Revision: D19862459

fbshipit-source-id: 462662630dcc17915558cdd8d144a5c6c3bfbefc