A bit of guard system code motion
Summary:
For profiling tracelets, we represent guards by a "guarded location", which is a pair:
1. An IR instruction that guards that location (local, stack, or member base register) to some type
2. A GuardConstraint struct containing constraints that irgen has declared on the type

The GuardConstraint is equivalent to a simple DataTypeCategory for all categories except DataTypeSpecialized. For this category, it can additionally contain information like which class's properties we actually used (to allow us to guard to some base class rather than to an overly-specific derived class).

We don't use this extra information today. When we save guarded locations to our profile data, we only store the type and category. As a result, we always guard objects on class equality rather than to a subclass constraint. I'd like to lift this restriction for "vanilla" array-likes, so that in profiling tracelets, we can distinguish between "guarding to an ArrayKind" and the weaker "guarding to vanilla". This first diff just introduces a new "relaxToConstrant" helper and calls it in places where we have the constraint. We continue to use the "relaxType" helper in optimizing tracelets, because it's easier to do guard relaxation with the simpler (type, category) schema. We also eliminate the separate "optimizeGuards" pass used in region-tracelet - instead, we always relax these types when collecting guards the first time.

In this diff, we also tweak some of the implementation of GuardConstraint logic slightly in a way that should be equivalent but more future-proof (e.g. we use isKnownDataType rather than listing types).

Reviewed By: ricklavoie

Differential Revision: D19857169

fbshipit-source-id: e4a34e3665a81d1b797ef7c6dc9c46f0c72cf6db