Support recursive legacy marking
Summary:
"Marking" a dvarray allows us to migrate them to mostly behave like a vec or dict except for those few behaviors that observe the legacy marking bit. We want to support both recursive and shallow marking, because different Hack frameworks will have different migration needs.

In this diff, we modify the recursive provenance tagging code to support an arbitrary "mutation" that takes an array-like input, potentially mutates it, and returns the result. I opted to keep the object warning in the common code, because it is a general failure of the recursive traversal, but I moved the remainder of the mutation into this helper closure that gets passed in. Both of our mutations have some state: provenance tagging has the memoized source location, and marking has flags for whether we've raised notice for different kinds of marking failures. We can treat all of this state as closure captures.

Added a new unit test that confirms that we don't run into refcounting issues when marking deeply or shallowly in a variety of configurations. This test uses the "is_array_marked_legacy" function to test whether marking was applied correctly; other tests in the same directory check that the legacy marking leads to actual downstream behavior differences (primarily in serialization).

Reviewed By: jjgriego

Differential Revision: D19910182

fbshipit-source-id: caeae6876710ceab8ad58eaa7f263caef5d7f06e