Remove Tnothing decl type
Summary:
Simplify type representation in advance of moving to Rust. The `Tnothing` decl type serves little purpose, as it has exactly the same interpretation as `Tunion []`. In fact, `nothing` can be considered syntax for empty union.

Important: the order of the constructors in the definition of `Typing_defs_core.ty_` is critical. In particular, all polymorphic constructors must appear before `locl` only constructors in the definition, otherwise our tooling that generates Rust definitions produces incompatible definitions. Hence the moving of `Tunion` and `Tintersection` in this diff.

Reviewed By: Wilfred

Differential Revision: D19951715

fbshipit-source-id: 52fbb4abf3990e7627701dbd06c45a4d5c1e8b06