Hack: Fix lint message on switch types
Summary:
## The Problem �
If a user writes Hack code like this:
```
<?hh

enum MyEnum: int {
  X = 0;
  Y = 1;
}

function takes_nullable_enum(?MyEnum $m): void {
  switch ($m) {
    case null:
      echo "It was null";
  }
}
```
We report a lint error:
```
$ hh --lint x.php
Lint[5614] This case value has type null, which is incompatible with type ?MyEnum.
   --> x.php
 11 |     case null:
    |          ^^^^
```
This isn't true! null is compatible with a nullable value.

The real problem here is that switch statements use coercion, so we should tell the user about that instead.

Differential Revision: D19957580

fbshipit-source-id: 6a59ef801f77933a7adced68c7e4a02bc73763d3