Don't refine pointer types across uses
Summary:
For member ops, if we know the base's type, we assert that the MBase pointer is an lval to that type, allowing us to do type-safe member ops.

Type assertions on lvals are not transferable across uses, however - just because they're true at one point doesn't mean that we can apply the type info to other program points. For example, if we have t1:LvalToArr = AssertType<LvalToArr> t0:LvalToCell, that doesn't mean that we'll still have an LvalToArr at later places where we use t0 as well. refineTmps makes this assumption, which can introduce problems if we escalate member bases. It's easy to fix.

Reviewed By: markw65

Differential Revision: D19973166

fbshipit-source-id: cdcf0f85a14c4335edb81f55e9eccdb4c604b0de