Remove queryExprStr requirement from TrustedAutoloadMap
Summary:
While writing a test for TrustedAutoloadMap, I found that we require an `Autoload.Query` option in `.hhvmconfig.hdf`, even though we never use it. We'll always have this option set in Facebook WWW, but it still seems like an oversight we may want to fix.

Note that we already check if the query option is empty in `ext_watchman_autoload.cpp`:

https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/fbcode/hphp/facebook/extensions/watchman_autoload/ext_watchman_autoload.cpp?commit=d143a7ff557bf07af85df83f5a5c3de0f8555d50&lines=346-348

So this extra check in `autoload-handler.cpp` really just makes the autoloader configuration more convoluted and less flexible for no benefit.

Reviewed By: dneiter

Differential Revision: D20043802

fbshipit-source-id: 70d8587756d00a4b1720ac22e0105f7c9da7278f