Add read-only mode to the SQLite wrapper
Summary:
iwonbigbro is packaging WWW as a read-only SquashFS volume, with autoload information stored in a SQLite DB somewhere in the volume.

My current native-autoloading plan for this world is to initialize an `HPHP::TrustedAutoloadMap` which reads from the SQLite DB. All should be well.

Except... when building a read-only DB, we set it in `journal_mode = DELETE` in accordance with [the SQLite docs](https://www.sqlite.org/wal.html#read_only_databases):

> Even though it is possible to open a read-only WAL-mode database, it is good practice to converted to PRAGMA journal_mode=DELETE prior to burning an SQLite database image onto read-only media.

Currently the SQLite wrapper always sets WAL mode on initialization, which causes a write to the otherwise read-only DB. Instead, let's open the DB in read-only mode and verify that everything still works.

Reviewed By: dneiter

Differential Revision: D20043225

fbshipit-source-id: 2f03a58c21b86cc03b09fe461941b64e228d8e2b