Fix an hhbbc type refinement crash
Summary:
There's a long standing issue with hhbbc's type system... it can't represent the intersection of two unrelated interfaces, or of an interface and an unrelated class.

In the new test case:

 - on the first round of analysis, when compiling g, hhbbc doesn't know what h returns, so $c is typed as the union_of(Cell, D) = Cell.
 - on the second round, it knows that h returns I1, and that it returns its parameter, and that cc returns I1, so we get union_of(I1, D) = I1.
 - on the third round it knows that cc returns C, and so h(cc()) returns C, so union_of(C, D) = A.

But A is not more refined than I1, and we fail assertions both when updating the use vars of the closure, and when updating the return type of g (obviously only one can actually fail).

We've previously tried to avoid these issues by preventing the inferrence of interfaces; but thats less than ideal because we can dispatch methods more efficiently when we know the object is a particular interface.

Instead, this relaxes the asserts to allow inferring a new return type of an unrelated class (or TObj) when the current return type is an interface. In that case it doesn't update the return type (to ensure monotonicity).

Reviewed By: ricklavoie

Differential Revision: D20056392

fbshipit-source-id: 7d7b8ffc6564f7e06ff2952beab1ad6de87e0dc4