Use MInstrEffects in load-elim
Summary:
MInstrEffects produces type bounds for the member base after a member op, given the op and the type of the base before it runs. We use this information in FrameState, but we don't use it in load-elim, losing type information for these ops. In this diff, we improve load-elim support for member instructions. We need to use load-elim to handle loops, since we pessimize them during initial irgen when the region is incomplete.

The effect is particularly pronounced in my "vanilla array-likes" prototype stack, because we want to guard to vanilla once and propagate that bit through the rest of the region. Without this diff, we doing O(n) vanilla checks for simple things like Vec\map. Aside from these vanilla checks, HHBBC mostly takes care of this analysis by inserting asserts.

Reviewed By: ricklavoie

Differential Revision: D20058191

fbshipit-source-id: 9692da2ef7f4905b08e45d9a464f34b905ea1e81