Do not use callee's ActRec in prepareUnpackArgs()
Summary:
This function will eventually run before ActRec is written, so it should not be
manipulating it in any way. It did so only for the purpose of updating the
number of arguments so that doFCall() can count the number of passed inout
arguments when unwinding, which may be different from the number of inout
arguments that the function takes. However, we already know the number of
originally passed arguments in doFCall(), so let's just use that.

Reviewed By: ricklavoie

Differential Revision: D20007939

fbshipit-source-id: 1d4e2e89ba7d11378cefaaa6b0a1a7c9f90129ca