Local-memory fixme provider for Local_memory and Decl_service backends
Summary: Store the HH_FIXMEs we collected during parsing in the Provider_backend.t, when using the Local_memory and Decl_service backends. I think this should work fine for sIDE. It may cause issues for the Decl_service if we still have cases where decl-FIXMEs in another file can suppress errors in the file being typechecked (and in that case, we'd need to send back FIXMEs from the decl service), but hopefully we'll eliminate those cases soon.

Reviewed By: arxanas

Differential Revision: D19819526

fbshipit-source-id: 5b4d823a732666874944bb084a829e773f1cde21