Make `namespace` a reserved keyword consistent with `function`
Summary:
Namespace is allowed as a name in locations that shouldn't work
```
class namespace {}
```
```
new namespace();
```
- Can't actually use the imported name since `namespace\` has special meaning

This diff bans this usage as parse errors by making `namespace` a reserved keyword.

`namespace` is only allowed as a name in the following circumstance: it is the prefix to a qualified name: This means that the `namespace` prefix will get elaborated as the current namespace
```
namespace\Foospace\bar();
```
and the prefix is not preceded with a backslash, i.e.
```
\namespace\Foospace\bar();
```
will result in an error.

 ---

Mild consequence:

namespace::$foo will be a runtime fatal parse error instead of a runtime fatal. Not that it makes _that_ big of a difference.

Reviewed By: Wilfred

Differential Revision: D19978598

fbshipit-source-id: 35c0d85e7837c87cb7931f2e80f76a889a91684e