shallow decls evict LazyLinearizations once complete
Summary:
During typechecking, we get "LazyLinearizations" on the local heap. These are linearizations-in-progress, ones that haven't yet been completed, and whose storage include closures for lazily computing the next element in the sequence. They are stored in a LocalCache of capacity 1000. If we need to work on a further decl when the LocalCache is already full, then we evict an existing one. Other lazy decls/linearizations might retain a reference to an evicted LazyLinearization, but any code which needs to look up a LazyLinearization by name will have to start a new one.

Whenever we push a linearization through to its end, it gets stored in the completed "Linearizations" heap in shared-memory. There's also a 1000-item-capacity local cache of this shared-mem object.

Whenever someone wants to look up a linearization by name, they first look in the completed "Linearizations" heap (using the LocalCache if it's already there, or copying+unmarshalling from shared-mem if it's not). If not in the Linearizations heap then they look in the LazyLinearizations heap. And if not there, then they start a new LazyLinearization.

I got some real-world numbers. For a moderately complex file, typechecking it produced 8000 LazyLinearizations and 6000 completed Linearizations.

I noticed that if a LazyLinearization has been completed, then it will be copide into the completed Linearizations heap, but it will retain its cache entry in the LazyLinearizations heap too. And it will quite possibly never be evicted, if it's accessed frequently enough.

So of the 8000 LazyLinearizations we used, 6000 of them were completed and written to the Linearizations heap, but how many of them were retained in the LazyLinearizations heap with its 1000-element capacity even though they'd never be looked up again? How many of them took up space, resulting in other ones being evicted?

This diff merely evicts a LazyLinearization from its local heap once it gets completed and copied into the completed Linearizations heap.

I expect this to maybe have a minor perf benefit, due to allowing the LazyLinearization heap to be better used.

Reviewed By: arxanas

Differential Revision: D20014435

fbshipit-source-id: f58a7a9d3688aa514161d775c6764390bec6ceb5