Make declaring an enum after use import a parse error
Summary:
Currently, the following produces an error:
```
use type NS\Foo;

class Foo {}
```
because we are trying to redeclare a class that has the same name as one that we've imported.

However, the following does not, even though one might expect it to:

```
use type NS\Foo;

enum Foo: int {}
```

 ---

I've changed the error message to be more generic, removing the "class" from the error message. First because enums seem slightly non class like (even though they are under the hood) and second, so I could reuse the error message, which I feel is descriptive enough for classes and functions, etc.

If there are opinions about this, let me know and I can make a specified error for each type.

Reviewed By: Wilfred, Matt-Schellhas

Differential Revision: D20013866

fbshipit-source-id: f29c53e1c4a3de4417e6455a676fe455b3bd9123