Optimize unwinder returning from an async function
Summary:
Prior to this diff, when the unwinder wanted to resume execution by returning a FSWH from an async function, it would call the caller's catch trace and then call resumeHelper to resume but that is inefficient and with jitted stack teardown, it will be incorrect.
This diff makes it so that we can just push the savedRip and do a asm ret when we are retuning to a jitted frame or still call resume helper when we are returning back to the interpreter since there is no catch trace.

Reviewed By: jano

Differential Revision: D20020027

fbshipit-source-id: 3652fb4bf474053c357138e4ba967d4a82836b31