Raise too many args if unpacking too many args
Summary:
prepareFuncEntry() was not observing too many arguments condition when a callee
was called with variadic arguments. Fix this by unifying the state observed by
prepareFuncEntry(). We no longer have untrimmed and trimmed state. The new
expectations are that there are up to func->numNonVariadicParams() arguments on
the stack and all other arguments are packed into a varray that can't be empty.
This matches JIT prologue behavior (except of emptiness).

Reviewed By: ricklavoie

Differential Revision: D20088967

fbshipit-source-id: 0f1ab96796d5e90ddea6267e6d0e495cdda6cb36