Add the target function/method as a dependency before adding implementation dependencies
Summary:
This is a bit subtle.  Consider the following example (added as a new test case below):
```
<<__Sealed(Mumble::class)>>
interface SealedInterface {
  public function method(): void;
  public function otherMethod(): void;
}

<<__Sealed(Mumble::class)>>
trait SealedTrait {
  require implements SealedInterface;
  public function method(): void {}
  public function otherMethod(): void {}
}

final class Mumble implements SealedInterface {
  use SealedTrait;
}
```
Currently, if we try to extract `SealedInterface::method`, we obtain:
```
<?hh
interface SealedInterface {
  public function method(): void;
}
trait SealedTrait {
  require implements \SealedInterface;
}
final class Mumble implements \SealedInterface {
  use \SealedTrait;
}
```
This is not well-typed: `Mumble` is a final class that is declared to implement `SealedInterface` but it doesn't provide an implementation of `method`.

First of all, why are `Mumble` and `SealedTrait` pulled into the extracted code?  Wouldn't it be sufficient to generate
```
interface SealedInterface {
  public function method(): void;
}
```
only?  It would, but in the process of type checking `SealedInterface` we in particular type check its user attributes, including `__Sealed`, which adds a dependency on `Mumble`, and that pulls in `SealedTrait`.  We could try and address this by clearing the user attributes before type checking the class containing the target method.  I haven't pursued this.

Instead, I tried to understand why the definitions of `Mumble` and/or `SealedTrait` lack `method`.  After all, a dependency on `SealedInterface::method` should have been added as an implementation dependency.  (We basically loop over all the pairs of class names in the current dependency set, and for each class-ancestor pair if we have a dependency on a method in the ancestor, we also add it as a dependency in the class.)

It turns out that we explicitly remove the dependency on `SealedInterface::method` from the set of dependency before closing it with respect to implementation dependencies.  And, in fact, in this case this dependency wasn't even added to the set in the first place because the target method is never called.  (It would be added if it were recursive and a call to it appeared in its body.)  So the remedy is to add the target method to the set of dependencies, add all implied implementation dependencies, and then remove the dependency on the target (so that we don't generate a stub for it; its body will be extracted verbatim).

Reviewed By: CatherineGasnier

Differential Revision: D20029481

fbshipit-source-id: 6028036db73d209f9e1f55b36a4541f39fb90dc0