Correctly generate initialisers from hints that are type definitions with generics
Summary:
Consider the following example (added as a new test case below):
```
type Id<T> = T;

class WithTypeAliasHint {
  private Id<int> $x = 0;

  public function getX(): Id<int> {
    return $this->x;
  }
}
```
The extraction of the `getX` method crashes because when we try to generate an initialiser for the property `$x` from its type `Id<int>`, we discover that it's a typedef, look up its definition, and try to construct an initialiser from the type the typedef aliases, but that type contains a generic type parameter and we don't propagate information about generics.  The simplest fix I can think of is to maintain a map from type parameter names to hints with which these type parameters are instantiated.  We already do this in a number of places in the type checker (e.g., `Typing_phase.localize`) but they are heavyweight and require passing in a typing environment.

Reviewed By: CatherineGasnier

Differential Revision: D20069894

fbshipit-source-id: 5e401dc2fd0d2fafff52a09240fd69f89fff6312