allow to infer partially annotated classes
Summary:
We can't easily create type vars for partially annotated types in the decl phase, because this would require fetching the decl for the class to check its type parameters.
Instead, we create the global type variables in Typing, with its ID being a hash of the position of the partial annotation.
We also fix the setting of the variances.

Differential Revision: D20036858

fbshipit-source-id: b49bdfb427e91b9df2ab3c29f6cbf7b2d0896aa2