Properly expand types in tasts
Summary:
Expand the types in the TAST before extracting globally inferred types.

When `build`-ing the position-to-type map, we need to make sure all local type variables are expanded. Otherwise, we might register a position-to-local-type-variable entry. Since we don't store the local type inference environment, we won't be able to expand this type when generating patches.

Reviewed By: CatherineGasnier

Differential Revision: D20116089

fbshipit-source-id: 20c4000da175b96ba26a358cad416cd308fa11a4