Fix crashes when doing "$x is Shape"
Summary:
"$x is Shape" means "check if $x is an object with class
literally named Shape", not that $x is a shape. The emitter was not
handling this corner case properly. It was treating it like a class
check (correctly), but still giving it a T_shape kind in the
type-structure. This leads to crashes during HHBBC or at runtime
because the type-structure has T_shape kind, but none of the other
fields we expect in a shape type-structure. A similar issue exists for
Tuple/tuples. Fix this by special casing this situation in the emitter
and treating it as "unresolved".

Reviewed By: oulgen

Differential Revision: D20125318

fbshipit-source-id: c2d6615877751a78d9f188f179f64e2df9624f1d