Tag enum names and values
Summary:
When array provenance is on, we'd like to give the arrays returned by Enum::getNames and Enum::getValues provenance based on the caller. However, these values are aggressively cached, which is quite important in idiomatic Hack code, as some enums may contain thousands of (statically-known) enum values. In particular, the caching layer converts persistent enum values into static arrays and non-persistent enum values into request-local cached arrays.

Tagging the largest of arrays on each call to the reflection methods is a no-go. Instead, we use a dummy "LargeEnum" tag which includes the enum class name for all enums above a configurable size limit, and we tag smaller enums precisely.

Reviewed By: jjgriego

Differential Revision: D20107209

fbshipit-source-id: ea089f249ce77d52f2f5ef2aca8d24ec7ad768c6