avoid waking up excessive threads upon improved health level
Summary:
We want to avoid waking up excessive threads when there is a long queue (because of not dequeuing under NoMore/BackOff). Increasing active threads too fast could put us back to an unhealthy state.
As a heuristic, stop introducing new workers if half of the threads are already active.  If we need to turn it off, we can set the Cautious threshold to the same as NoMore threshold, so we won't enter the cautious state.

Reviewed By: markw65

Differential Revision: D20137207

fbshipit-source-id: 03459b3b980c771a6f942e7e305eefe9e051799d