Mark ops with vanilla inputs and outputs
Summary:
In order to support hidden classes for array-likes, we have to distinguish between IR ops that only operate on array-likes with standard "vanilla" layouts and ones that can also operate on array-likes with "bespoke" hidden-class layout. "TArr", "TVec", etc. refer to the Hack data type and NOT to the runtime layout; we'll use "TVanillaArr", "TVanillaVec", etc. to refer to the types of array-likes with the standard layout. Of our current IR ops, all but a small handful require vanilla inputs. We can find these 11 "layout-agnostic" ops with grep:
```
$ grep -P '^\|' hphp/doc/ir.specification | grep '[DS]([^ ]*\(Arr\|Vec\|Dict\|Keyset\)' | grep -v Vanilla
| CheckVArray, DArrPacked, S(Arr), B|P
| CheckDArray, DArrMixed, S(Arr), B|P
| CheckDVArray, D(Arr), S(Arr), B|P
| EqArrayDataPtr,              D(Bool), S(ArrLike) S(ArrLike),            NF
| RaiseHackArrParamNotice<type,id,isReturn>, ND, S(Arr) S(Func), NF
| RaiseHackArrPropNotice<type>, ND, S(Cls) S(Arr) S(Int) C(Bool), NF
| RaiseArraySerializeNotice, ND, C(Int) S(ArrLike), NF
| RaiseTooManyArg<func>, ND, S(Arr,Vec), CRc
| CheckArrayCOW, ND, S(ArrLike), B
| ThrowOutOfBounds, ND, S(ArrLike|Obj) S(Cell), T
| ThrowInvalidArrayKey, ND, S(ArrLike) S(Cell), T
```

We decided that it's better to make these annotations explicit rather than implicit. In this diff, we simply annotate ops that require vanilla inputs. For now, TVanillaArr is just TArr - we'll change that in the next diff.

Reviewed By: ricklavoie

Differential Revision: D20148947

fbshipit-source-id: 3a4e13242111cf21476a6e02eafc2234e235f050