Add vanilla types to jit::Type
Summary:
This diff adds a "vanilla" bit to the ArraySpec, which is set for array-likes that are known to have the default layout for their type. For TVanillaArr, that means having one of the existing ArrayKind layouts; for TVanillaVec, TVanillaDict, and TVanillaKeyset, there is exactly one vanilla layout for the type. Note that right now, ALL array-likes have a vanilla type; the alternative "bespoke" array-like types are completely uninhabited. Nevertheless, we can use the JIT type system to verify that we're correctly checking for vanilla layouts.

The interaction of these types with HHBBC's RepoAuthType requires some explanation. Since our goal is to create bespoke types on the fly, in the runtime, we never tell HHBBC about these types. As a result, when HHBBC provides us the RepoAuthType "VArr:N(Int,Str)" for a given array-like, what we learn is that "if the array is vanilla, then it is a two-element packed array with an int value and a string value". We represent this within ArraySpec by making the kind, type, and vanilla specialization dimensions completely orthogonal, but if the vanilla bit is unset, the kind dimension is inaccessible. This logic is tested in new type unit tests.

One more note about naming. We'd like to avoid cluttering debug information with excessive discussion of "bespoke" and "vanilla". As a result, when we have both the kind and vanilla bit for a TArr, we'll print the type as "TArr=PackedKind", as we do today. If we don't have the vanilla bit, we'll instead print the type as "TArr={PackedKind|Bespoke}" - this type is rarer and so more verbose than the other type. If we only have the vanilla bit, we'll print "TArr=Vanilla".

This diff is a no-op as long as AllowBespokeArrayLikes is false; at strategic places, we check this flag and ensure that we're not pessimizing any types. In the next diff, we'll move the usage of this flag to a central location that actually generates AssertType<TVanillaArrLike> / CheckType<TVanillaArrLike>.

Reviewed By: ricklavoie

Differential Revision: D19819483

fbshipit-source-id: 4cb0a1a3121e3b78a44b059306ac0718c80366ea